% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimReduce.R
\name{dimReduce}
\alias{dimReduce}
\title{Get lower dimension embedding}
\usage{
dimReduce(x, flavor = c("pca", "tsne", "umap"), k = 2, is.counts = TRUE,
  ntop = 500)
}
\arguments{
\item{x}{gene expresison matrix [GENES x SAMPLES]}

\item{flavor}{the algorithm to use to obtain the dimensionality reduction
must be in c('pca', 'tsne', 'umap')}

\item{k}{the number of dimensions in the reduced dimension representation}

\item{is.counts}{logical: is `x` counts data}

\item{ntop}{number of most variable genes to use for dimensionality
reduction}
}
\value{
reduced dimensionality representation
}
\description{
Get lower dimension embedding
}
\keyword{internal}
