% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human.ppi}
\alias{human.ppi}
\title{Human Protein-Protein interaction graph}
\format{A square matrix where A_{ij}=1 if gene i interacts with gene j}
\source{
\url{http://www.string-db.org/}
}
\usage{
human.ppi
}
\description{
An adjacency matrix of the 10 percent highest confidence interactions
between human proteins on STRINGdb.
}
\details{
See the script in `system.file(package="netSmooth", "data-raw", "make_ppi_from_string.R")`
for full details of how this object was made.
}
\keyword{datasets}
