% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{l1.normalize.columns}
\alias{l1.normalize.columns}
\title{Column-normalize a sparse, symmetric matrix (using the l1 norm) so that each
column sums to 1.}
\usage{
l1.normalize.columns(A)
}
\arguments{
\item{A}{matrix}
}
\value{
row-normalized sparse matrix object
}
\description{
Column-normalize a sparse, symmetric matrix (using the l1 norm) so that each
column sums to 1.
}
