% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{l1.normalize.rows}
\alias{l1.normalize.rows}
\title{Row-normalize a sparse matrix (using the l1 norm) so that each
row sums to 1.}
\usage{
l1.normalize.rows(A)
}
\arguments{
\item{A}{matrix}
}
\value{
row-normalized sparse matrix object
}
\description{
Row-normalize a sparse matrix (using the l1 norm) so that each
row sums to 1.
}
