% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse.ppi}
\alias{mouse.ppi}
\title{Mouse Protein-Protein interaction graph}
\format{A square matrix where A_{ij}=1 if gene i interacts with gene j}
\source{
\url{http://www.string-db.org/}
}
\usage{
mouse.ppi
}
\description{
An adjacency matrix of the 10 percent highest confidence interactions
between mouse proteins on STRINGdb
}
\keyword{datasets}
