% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectFromNetworkRecombine.R
\name{projectFromNetworkRecombine}
\alias{projectFromNetworkRecombine}
\title{Combine gene expression from smoothed space (that of the network) with the
expression of genes that were not smoothed (not present in network)}
\usage{
projectFromNetworkRecombine(original_expression, smoothed_expression)
}
\arguments{
\item{original_expression}{the non-smoothed expression}

\item{smoothed_expression}{the smoothed gene expression, in the space
of the genes defined by the network}
}
\value{
a matrix in the dimensions of original_expression, where values that
        are present in smoothed_expression are copied from there.
}
\description{
Combine gene expression from smoothed space (that of the network) with the
expression of genes that were not smoothed (not present in network)
}
\keyword{internal}
