% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{random.walk.by.diffusr}
\alias{random.walk.by.diffusr}
\title{Smooth data on graph by performing random walk iterations until convergence
using Cpp code from the diffusr package.}
\usage{
random.walk.by.diffusr(gene_expression, adj_matrix, lambda)
}
\arguments{
\item{f0}{initial data matrix [NxM]}

\item{lambda}{smoothing coefficient (1 - restart probability of
random walk)}

\item{adj_matrix}{adjacency matrix of graph to network smooth on
will be column-normalized.}
}
\value{
network-smoothed gene expression
}
\description{
Smooth data on graph by performing random walk iterations until convergence
using Cpp code from the diffusr package.
}
