% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{random.walk.by.iterations}
\alias{random.walk.by.iterations}
\title{Smooth data on graph by performing random walk iterations until convergence}
\usage{
random.walk.by.iterations(gene_expression, adj_matrix, lambda)
}
\arguments{
\item{f0}{initial data matrix [NxM]}

\item{lambda}{smoothing coefficient (1 - restart probability of
random walk)}

\item{max.iter}{maximum number of iterations to allow for convergence.
if not converfed in max.iter tries, gives warning.}

\item{tol}{error tolerance for convergence criterion:
will terminate when norm(f_t - f_{t+1}) < tol}

\item{adj_matrix}{adjacency matrix of graph to network smooth on
will be column-normalized.}
}
\value{
network-smoothed gene expression
}
\description{
Smooth data on graph by performing random walk iterations until convergence
}
