% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomWalkByIterations.R
\name{randomWalkByIterations}
\alias{randomWalkByIterations}
\title{Smooth data on graph by computing iterations}
\usage{
randomWalkByIterations(f0, adjMatrix, alpha, normalizeAjdMatrix = c("rows",
  "columns"), tol = 1e-06, max.iter = 100)
}
\arguments{
\item{f0}{initial data matrix [NxM]}

\item{adjMatrix}{adjacency matrix of graph to network smooth on
will be column-normalized.}

\item{alpha}{smoothing coefficient (1 - restart probability of
random walk)}

\item{tol}{the tolerance (stopping criterion)}

\item{max.iter}{the maximum number of iterations before terminating}
}
\value{
network-smoothed gene expression
}
\description{
Smooth data on graph by computing iterations
}
\keyword{internal}
