% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomWalkBySolve.R
\name{randomWalkBySolve}
\alias{randomWalkBySolve}
\title{Smooth data on graph by solving the linear equation
(I - alpha*A)^T * E_sm^T = E^T * (1-alpha)}
\usage{
randomWalkBySolve(E, A, alpha, normalizeAjdMatrix = c("rows", "columns"))
}
\arguments{
\item{E}{initial data matrix [NxM]}

\item{A}{adjacency matrix of graph to network smooth on
will be column-normalized.}

\item{alpha}{smoothing coefficient (1 - restart probability of
random walk)}
}
\value{
network-smoothed gene expression
}
\description{
Smooth data on graph by solving the linear equation
(I - alpha*A)^T * E_sm^T = E^T * (1-alpha)
}
\keyword{internal}
