% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{robustClusters}
\alias{robustClusters}
\title{Perform robust clustering on dataset, and calculate the proportion of
samples in robust clusters}
\usage{
robustClusters(se, dimReduceFlavor = "auto", is.counts = TRUE, ...)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{dimReduceFlavor}{algorithm for dimensionality reduction step
of clustering procedure. May be 'pca', 'tsne',
'dm' or 'auto', which uses shannon entropy to
pick the algorithm.}

\item{is.counts}{logical: is the data counts}

\item{...}{arguments passed on to `clusterExperimentWorkflow`}
}
\value{
list(clusters, proportion.robust)
}
\description{
Perform robust clustering on dataset, and calculate the proportion of
samples in robust clusters
}
