% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smallPPI}
\alias{smallPPI}
\title{A small human Protein-Protein interaction graph for use in examples.}
\format{An object of class \code{matrix} with 611 rows and 611 columns.}
\usage{
smallPPI
}
\description{
Contains a synthetic PPI of human genes.
}
\keyword{datasets}
