% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smallscRNAseq}
\alias{smallscRNAseq}
\title{A small single cell RNA-seq dataset for use in examples.}
\format{SingleCellExperiment}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE44183}
}
\usage{
smallscRNAseq
}
\description{
Contains scRNAseq profiles of human blastomeres.
}
\keyword{datasets}
