% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimreduce.R
\name{calc2DEntropy}
\alias{calc2DEntropy}
\title{Calculate entropy in 2D data}
\usage{
calc2DEntropy(x, numBins1 = 20, numBins2 = 20)
}
\arguments{
\item{x}{the 2D data to get entropy from}

\item{numBins1}{the number of bins along the first dimension to
discretize data into}

\item{numBins2}{the number of bins along the second dimension to
discretize data into}
}
\value{
The Shannon entropy in the 2D data x
}
\description{
Calculate entropy in 2D data
}
