% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterOne.R
\name{clusterOne}
\alias{clusterOne}
\title{Run one clustering using kmeans o PAM}
\usage{
clusterOne(x, algorithm = c("kmeans", "pam"), k = 5)
}
\description{
Run one clustering using kmeans o PAM
}
\keyword{internal}
