% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimreduce.R
\name{dimReduce}
\alias{dimReduce}
\title{Get lower dimension embedding}
\usage{
dimReduce(expr, flavor = c("pca", "tsne"), k = 2, is.counts = TRUE)
}
\arguments{
\item{expr}{gene expresison matrix [GENES x SAMPLES]}

\item{flavor}{the algorithm to use to obtain the dimensionality reduction
must be in c('pca', 'tsne')}

\item{k}{the number of dimensions in the reduced dimension representation}

\item{is.counts}{logical: is `expr` counts data}
}
\value{
reduced dimensionality representation
}
\description{
Get lower dimension embedding
}
