% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netsmooth.R
\name{netSmooth}
\alias{netSmooth}
\title{Perform network smoothing of gene expression or other omics data}
\usage{
netSmooth(x, adjMatrix, alpha = "optimal",
  optimalAlphaMethod = c("robustness", "entropy"), optimalAlphaRange = 0.1 *
  (1:9), summarizedExprrimentAssay = 1, is.counts = TRUE, ...)
}
\arguments{
\item{x}{matrix or SummarizedExperiment}

\item{adjMatrix}{adjacency matrix of gene network to use}

\item{alpha}{numeric in [0,1] or 'optimal'. if 'optimal', the optimal
value for alpha will be automatically chosen among the values
specified in `optimalAlphaRange`, using the strategy
specified in `optimalAlphaMethod`}

\item{optimalAlphaMethod}{if 'robustness', pick alpha that gives the
highest proportion of samples in robust clusters
if 'entropy', pick alpha that gives highest
Shannon entropy in 2D PCA embedding}

\item{optimalAlphaRange}{if `alpha='optimal'`, search these values
for the best alpha}

\item{is.counts}{logical: is the assay count data}

\item{...}{arguments passed on to `robustClusters` if using the robustness
criterion for optimizing alpha}

\item{summarizedExperimentAssay}{if `x` is a SummarizedExperiment object,
the index of the assay to use}
}
\value{
network-smoothed gene expression matrix or SummarizedExperiment
        object
}
\description{
Perform network smoothing of gene expression or other omics data
}
