% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netSmooth.R
\name{netSmooth,matrix-method}
\alias{netSmooth,matrix-method}
\alias{netSmooth}
\alias{netSmooth,SingleCellExperiment-method}
\alias{netSmooth,Matrix-method}
\alias{netSmooth,DelayedMatrix-method}
\title{Perform network smoothing of gene expression or other omics data}
\usage{
\S4method{netSmooth}{matrix}(x, adjMatrix, alpha = 0.5, normalizeAdjMatrix = c("rows", "columns"))

\S4method{netSmooth}{SingleCellExperiment}(x, adjMatrix, alpha = 0.5, normalizeAdjMatrix = c("rows", "columns"))

\S4method{netSmooth}{Matrix}(x, adjMatrix, alpha = 0.5, normalizeAdjMatrix = c("rows", "columns"))

\S4method{netSmooth}{DelayedMatrix}(
  x,
  adjMatrix,
  alpha = 0.5,
  normalizeAdjMatrix = c("rows", "columns"),
  filepath = NULL
)
}
\arguments{
\item{x}{single cell Experiment}

\item{adjMatrix}{adjacency matrix of gene network to use}

\item{alpha}{numeric in [0,1]}

\item{normalizeAdjMatrix}{how to normalize the adjacency matrix
possible values are 'rows' (in-degree)
and 'columns' (out-degree)}

\item{filepath}{String: Path to location where hdf5 output file is supposed to be saved.
Will be ignored when regular matrices or SummarizedExperiment are
used as input.}
}
\value{
network-smoothed gene expression matrix or SummarizedExperiment
        object
}
\description{
Perform network smoothing of gene expression or other omics data
}
\examples{
x <- matrix(rnbinom(12000, size=1, prob = .1), ncol=60)
rownames(x) <- paste0('gene', seq_len(dim(x)[1]))

adj_matrix <- matrix(as.numeric(rnorm(200*200)>.8), ncol=200)
rownames(adj_matrix) <- colnames(adj_matrix) <- paste0('gene', seq_len(dim(x)[1]))
x.smoothed <- netSmooth(x, adj_matrix, alpha=0.5)
}
