% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimreduce.R
\name{pickDimReduction}
\alias{pickDimReduction}
\title{Pick the dimensionality reduction method for a dataset that gives the
2D embedding with the highest entropy}
\usage{
pickDimReduction(expr)
}
\arguments{
\item{x}{matrix or SummarizedExperiment object [GENES x SAMPLES]}

\item{flavors}{list of dimensionality reduction algorithms to try}

\item{is.counts}{logical: is exprs count data}
}
\value{
name of dimensionality reduction method that gives the highest
           2d entropy
}
\description{
Pick the dimensionality reduction method for a dataset that gives the
2D embedding with the highest entropy
}
\examples{
x <- matrix(rnbinom(60000, size=1, prob = .1), ncol=100)
pickDimReduction(x)
}
