% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{robustClusters}
\alias{robustClusters}
\title{Perform robust clustering on dataset, and calculate the proportion of
samples in robust clusters}
\usage{
robustClusters(se, ...)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{...}{arguments passed on to `clusterExperimentWorkflow`}
}
\value{
list(clusters, proportion.robust)
}
\description{
Perform robust clustering on dataset, and calculate the proportion of
samples in robust clusters
}
